Move logging out of the config module.
Also corrects a mistake in the config-pom that referenced the old packages.
