+++ pom.xml
@@ -15,6 +15,7 @@
 		<module>jaxrs-api</module>
 		<module>jaxrs</module>
 		<module>config</module>
+		<module>logging</module>
 		<module>examples</module>
 	</modules>
